Retrieve updated cellxgene database metadata

Retrieve updated cellxgene database metadata

Usage

db(overwrite = .db_online() && .db_first())

Arguments

overwrite

logical(1) indicating whether the database of collections should be updated from the internet (the default, when internet is available and, in an interactive session, the user requests the update), or read from disk (assuming previous successful access to the internet). overwrite = FALSE might be useful for reproducibility, testing, or when working in an environment with restricted internet access.

Value

db() returns an object of class 'cellxgene_db', summarizing available collections, datasets, and files.

Details

The database is retrieved from the cellxgene data portal web site. 'collections' metadata are retrieved on each call; metadata on each collection is cached locally for re-use.

Examples

db()
#> Collections ■                                  1% | ETA:  1m
#> Collections ■■                                 3% | ETA:  2m
#> Collections ■■■                                6% | ETA:  2m
#> Collections ■■■■                              11% | ETA:  1m
#> Collections ■■■■■■                            16% | ETA:  1m
#> Collections ■■■■■■■                           20% | ETA:  1m
#> Collections ■■■■■■■■                          24% | ETA:  1m
#> Collections ■■■■■■■■■                         28% | ETA:  1m
#> Collections ■■■■■■■■■■                        31% | ETA:  1m
#> Collections ■■■■■■■■■■■                       32% | ETA:  1m
#> Collections ■■■■■■■■■■■                       34% | ETA:  1m
#> Collections ■■■■■■■■■■■■■                     39% | ETA:  1m
#> Collections ■■■■■■■■■■■■■■                    43% | ETA: 48s
#> Collections ■■■■■■■■■■■■■■■■                  48% | ETA: 42s
#> Collections ■■■■■■■■■■■■■■■■■                 54% | ETA: 36s
#> Collections ■■■■■■■■■■■■■■■■■■■               60% | ETA: 31s
#> Collections ■■■■■■■■■■■■■■■■■■■■■             66% | ETA: 25s
#> Collections ■■■■■■■■■■■■■■■■■■■■■■            71% | ETA: 21s
#> Collections ■■■■■■■■■■■■■■■■■■■■■■■■          75% | ETA: 18s
#> Collections ■■■■■■■■■■■■■■■■■■■■■■■■■         80% | ETA: 14s
#> Collections ■■■■■■■■■■■■■■■■■■■■■■■■■■■       86% | ETA: 10s
#> Collections ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     92% | ETA:  5s
#> Collections ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    95% | ETA:  3s
#> Collections ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
#> cellxgene_db
#> number of collections(): 302
#> number of datasets(): 1831
#> number of files(): 1859