Facets available for querying cellxgene data

FACETS is a character vector of common fields used to subset cellxgene data.

facets() is used to query the cellxgene database for current values of one or all facets.

facets_filter() provides a convenient way to filter facets based on label or ontology term.

Usage

FACETS

facets(cellxgene_db = db(), facets = FACETS)

facets_filter(facet, key = c("label", "ontology_term_id"), value, exact = TRUE)

Format

FACETS is an object of class character of length 8.

Arguments

cellxgene_db

an (optional) cellxgene_db object, as returned by db().

facets

a character() vector corersponding to one of the facets in FACETS.

facet

the column containing faceted information, e.g., sex in datasets(db).

key

character(1) identifying whether value is a label or ontology_term_id.

value

character() value of the label or ontology term to filter on. The value may be a vector with length(value) > 0 for exact matchs (exact = TRUE, default), or a character(1) regular expression.

exact

logical(1) whether values match exactly (default, TRUE) or as a regular expression (FALSE).

Value

facets() returns a tibble with columns facet, label, ontology_term_id, and n, the number of times the facet label is used in the database.

facets_filter() returns a logical vector with length equal to the length (number of rows) of facet, with TRUE

indicating that the value of key is present in the dataset.

Examples

f <- facets()

## levels of each facet
f |>
    dplyr::count(facet)
#> # A tibble: 8 × 2
#>   facet                       n
#>   <chr>                   <int>
#> 1 assay                      34
#> 2 cell_type                 682
#> 3 development_stage         215
#> 4 disease                    80
#> 5 organism                    5
#> 6 self_reported_ethnicity    28
#> 7 sex                         3
#> 8 tissue                    302

## same as facets(, facets = "organism")
f |>
    dplyr::filter(facet == "organism")
#> # A tibble: 5 × 4
#>   facet    label                 ontology_term_id     n
#>   <chr>    <chr>                 <chr>            <int>
#> 1 organism Homo sapiens          NCBITaxon:9606     657
#> 2 organism Mus musculus          NCBITaxon:10090    188
#> 3 organism Callithrix jacchus    NCBITaxon:9483       4
#> 4 organism Sus scrofa domesticus NCBITaxon:9825       3
#> 5 organism Macaca mulatta        NCBITaxon:9544       2

db <- db()
ds <- datasets(db)

## datasets with African American females
ds |>
    dplyr::filter(
        facets_filter(self_reported_ethnicity, "label", "African American"),
        facets_filter(sex, "label", "female")
    )
#> # A tibble: 45 × 26
#>    dataset_id                 collection_id donor_id assay  cell_count cell_type
#>    <chr>                      <chr>         <list>   <list>      <int> <list>   
#>  1 8693bc2c-d84f-450f-983a-d… bcb61471-2a4… <list>   <list>     304652 <list>   
#>  2 37f30005-8a5c-4bdf-88aa-a… bcb61471-2a4… <list>   <list>     172847 <list>   
#>  3 54914718-c0d3-4e17-9e06-d… bcb61471-2a4… <list>   <list>     107344 <list>   
#>  4 39c0575c-5f9c-4009-bac0-6… 7d7cabfd-1d1… <list>   <list>     600929 <list>   
#>  5 4c4dedb7-1f74-4be7-9916-2… b953c942-f5d… <list>   <list>      23197 <list>   
#>  6 8716534a-4c1c-414c-aa98-4… b953c942-f5d… <list>   <list>      96129 <list>   
#>  7 24205601-0780-4bf2-b1d9-0… c9706a92-0e5… <list>   <list>      31696 <list>   
#>  8 a963dd1b-ee7e-42b3-8afe-0… a98b828a-622… <list>   <list>      48783 <list>   
#>  9 b976e4cb-968a-4ee2-b11d-b… 436154da-bcf… <list>   <list>    1263676 <list>   
#> 10 0899734f-df71-41a6-a1ea-c… a238e9fa-2bd… <list>   <list>       2541 <list>   
#> # ℹ 35 more rows
#> # ℹ 20 more variables: dataset_deployments <chr>, development_stage <list>,
#> #   disease <list>, is_primary_data <chr>, is_valid <lgl>,
#> #   mean_genes_per_cell <dbl>, name <chr>, organism <list>,
#> #   processing_status <list>, published <lgl>, revision <int>,
#> #   schema_version <chr>, self_reported_ethnicity <list>, sex <list>,
#> #   suspension_type <list>, tissue <list>, tombstone <lgl>, …

## datasets with non-European, known ethnicity
facets(db, "self_reported_ethnicity")
#> # A tibble: 28 × 4
#>    facet                   label                          ontology_term_id     n
#>    <chr>                   <chr>                          <chr>            <int>
#>  1 self_reported_ethnicity European                       HANCESTRO:0005     402
#>  2 self_reported_ethnicity unknown                        unknown            289
#>  3 self_reported_ethnicity na                             na                 190
#>  4 self_reported_ethnicity Asian                          HANCESTRO:0008     102
#>  5 self_reported_ethnicity African American               HANCESTRO:0568      54
#>  6 self_reported_ethnicity Hispanic or Latin American     HANCESTRO:0014      40
#>  7 self_reported_ethnicity African American or Afro-Cari… HANCESTRO:0016      26
#>  8 self_reported_ethnicity multiethnic                    multiethnic         25
#>  9 self_reported_ethnicity Greater Middle Eastern  (Midd… HANCESTRO:0015      22
#> 10 self_reported_ethnicity British                        HANCESTRO:0462      10
#> # ℹ 18 more rows
ds |>
    dplyr::filter(
        !facets_filter(
            self_reported_ethnicity, "label", c("European", "na", "unknown")
        )
    )
#> # A tibble: 22 × 26
#>    dataset_id                 collection_id donor_id assay  cell_count cell_type
#>    <chr>                      <chr>         <list>   <list>      <int> <list>   
#>  1 a8c0b8af-0d67-4b93-ac2d-4… 9c8808ce-113… <list>   <list>       4355 <list>   
#>  2 d374b984-a6e0-4d47-ba6d-e… 8e880741-bf9… <list>   <list>      17704 <list>   
#>  3 e567706a-6f68-4789-ad14-2… 8e880741-bf9… <list>   <list>      25741 <list>   
#>  4 72ad5ea4-1cbe-4862-a03e-f… 8e880741-bf9… <list>   <list>      17777 <list>   
#>  5 0f1098a9-35d4-427f-9040-7… 8e880741-bf9… <list>   <list>      23857 <list>   
#>  6 4116ae06-9e4c-4f8d-9b26-4… 9d63fcf1-5ca… <list>   <list>       3596 <list>   
#>  7 57d85cba-528e-45e0-9817-e… 9d63fcf1-5ca… <list>   <list>      10734 <list>   
#>  8 119ce3e3-c438-4bae-9a45-2… 9d63fcf1-5ca… <list>   <list>       4880 <list>   
#>  9 84fa33d5-8089-417e-9efe-e… 9d63fcf1-5ca… <list>   <list>       3851 <list>   
#> 10 ea0ec660-2262-46d8-b785-2… 9d63fcf1-5ca… <list>   <list>      45528 <list>   
#> # ℹ 12 more rows
#> # ℹ 20 more variables: dataset_deployments <chr>, development_stage <list>,
#> #   disease <list>, is_primary_data <chr>, is_valid <lgl>,
#> #   mean_genes_per_cell <dbl>, name <chr>, organism <list>,
#> #   processing_status <list>, published <lgl>, revision <int>,
#> #   schema_version <chr>, self_reported_ethnicity <list>, sex <list>,
#> #   suspension_type <list>, tissue <list>, tombstone <lgl>, …