Facets available for querying cellxgene data

FACETS is a character vector of common fields used to subset cellxgene data.

facets() is used to query the cellxgene database for current values of one or all facets.

facets_filter() provides a convenient way to filter facets based on label or ontology term.

Usage

FACETS

facets(cellxgene_db = db(), facets = FACETS)

facets_filter(facet, key = c("label", "ontology_term_id"), value, exact = TRUE)

Format

FACETS is an object of class character of length 8.

Arguments

cellxgene_db

an (optional) cellxgene_db object, as returned by db().

facets

a character() vector corersponding to one of the facets in FACETS.

facet

the column containing faceted information, e.g., sex in datasets(db).

key

character(1) identifying whether value is a label or ontology_term_id.

value

character() value of the label or ontology term to filter on. The value may be a vector with length(value) > 0 for exact matchs (exact = TRUE, default), or a character(1) regular expression.

exact

logical(1) whether values match exactly (default, TRUE) or as a regular expression (FALSE).

Value

facets() returns a tibble with columns facet, label, ontology_term_id, and n, the number of times the facet label is used in the database.

facets_filter() returns a logical vector with length equal to the length (number of rows) of facet, with TRUE

indicating that the value of key is present in the dataset.

Examples

f <- facets()

## levels of each facet
f |>
    dplyr::count(facet)
#> # A tibble: 8 × 2
#>   facet                       n
#>   <chr>                   <int>
#> 1 assay                      33
#> 2 cell_type                 703
#> 3 development_stage         216
#> 4 disease                    84
#> 5 organism                    5
#> 6 self_reported_ethnicity    30
#> 7 sex                         3
#> 8 tissue                    330

## same as facets(, facets = "organism")
f |>
    dplyr::filter(facet == "organism")
#> # A tibble: 5 × 4
#>   facet    label                 ontology_term_id     n
#>   <chr>    <chr>                 <chr>            <int>
#> 1 organism Homo sapiens          NCBITaxon:9606     704
#> 2 organism Mus musculus          NCBITaxon:10090    188
#> 3 organism Callithrix jacchus    NCBITaxon:9483      26
#> 4 organism Sus scrofa domesticus NCBITaxon:9825       3
#> 5 organism Macaca mulatta        NCBITaxon:9544       2

db <- db()
ds <- datasets(db)

## datasets with African American females
ds |>
    dplyr::filter(
        facets_filter(self_reported_ethnicity, "label", "African American"),
        facets_filter(sex, "label", "female")
    )
#> # A tibble: 48 × 24
#>    dataset_id   dataset_version_id collection_id donor_id assay  batch_condition
#>    <chr>        <chr>              <chr>         <list>   <list> <list>         
#>  1 218acb0f-9f… cc4284ca-8118-4b1… 436154da-bcf… <list>   <list> <lgl [1]>      
#>  2 e47c65a8-7d… 27b49e7c-0d90-42c… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  3 c8d40d53-38… 48450fdc-ff62-498… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  4 c45f05b3-aa… d71044a8-2326-4ec… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  5 a6388a6f-60… 3cc5cb70-7442-4f5… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  6 a41202e6-17… 38b93dd0-5edd-47a… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  7 87183784-5b… ad925422-3011-470… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  8 842c6f5d-4a… 29cf079b-e74c-415… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  9 7f08cbcc-57… cd74e95e-6583-487… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#> 10 74520626-b0… 4e74cb9d-f8f5-43a… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#> # ℹ 38 more rows
#> # ℹ 18 more variables: cell_count <int>, cell_type <list>,
#> #   development_stage <list>, disease <list>, explorer_url <chr>,
#> #   is_primary_data <list>, mean_genes_per_cell <dbl>, organism <list>,
#> #   schema_version <chr>, self_reported_ethnicity <list>, sex <list>,
#> #   suspension_type <list>, tissue <list>, title <chr>, tombstone <lgl>,
#> #   x_approximate_distribution <chr>, published_at <date>, revised_at <date>

## datasets with non-European, known ethnicity
facets(db, "self_reported_ethnicity")
#> # A tibble: 30 × 4
#>    facet                   label                          ontology_term_id     n
#>    <chr>                   <chr>                          <chr>            <int>
#>  1 self_reported_ethnicity European                       HANCESTRO:0005     431
#>  2 self_reported_ethnicity unknown                        unknown            311
#>  3 self_reported_ethnicity na                             na                 212
#>  4 self_reported_ethnicity Asian                          HANCESTRO:0008     130
#>  5 self_reported_ethnicity African American               HANCESTRO:0568      57
#>  6 self_reported_ethnicity Hispanic or Latin American     HANCESTRO:0014      41
#>  7 self_reported_ethnicity admixed ancestry               HANCESTRO:0306      28
#>  8 self_reported_ethnicity African American or Afro-Cari… HANCESTRO:0016      26
#>  9 self_reported_ethnicity multiethnic                    multiethnic         25
#> 10 self_reported_ethnicity Greater Middle Eastern  (Midd… HANCESTRO:0015      22
#> # ℹ 20 more rows
ds |>
    dplyr::filter(
        !facets_filter(
            self_reported_ethnicity, "label", c("European", "na", "unknown")
        )
    )
#> # A tibble: 22 × 24
#>    dataset_id   dataset_version_id collection_id donor_id assay  batch_condition
#>    <chr>        <chr>              <chr>         <list>   <list> <list>         
#>  1 e6a11140-25… b79122e0-7f34-4c9… e5f58829-1a6… <list>   <list> <lgl [1]>      
#>  2 6ec405bb-47… e9ec2d80-03ef-47c… e5f58829-1a6… <list>   <list> <lgl [1]>      
#>  3 2ba40233-85… bd265e27-eee1-49f… e5f58829-1a6… <list>   <list> <lgl [1]>      
#>  4 2423ce2c-31… caaae06a-80e1-480… e5f58829-1a6… <list>   <list> <lgl [1]>      
#>  5 c45f05b3-aa… d71044a8-2326-4ec… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  6 87183784-5b… ad925422-3011-470… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  7 6e96d93c-de… 37e6e34e-ef55-4d8… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  8 31b49393-44… d6f77d6a-6a1e-4bf… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  9 2d3e3c41-34… 93c62ade-711e-49a… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#> 10 cfa3c355-ee… 18f9edfa-8311-4d5… 9c8808ce-113… <list>   <list> <lgl [1]>      
#> # ℹ 12 more rows
#> # ℹ 18 more variables: cell_count <int>, cell_type <list>,
#> #   development_stage <list>, disease <list>, explorer_url <chr>,
#> #   is_primary_data <list>, mean_genes_per_cell <dbl>, organism <list>,
#> #   schema_version <chr>, self_reported_ethnicity <list>, sex <list>,
#> #   suspension_type <list>, tissue <list>, title <chr>, tombstone <lgl>,
#> #   x_approximate_distribution <chr>, published_at <date>, revised_at <date>