Facets available for querying cellxgene data

FACETS is a character vector of common fields used to subset cellxgene data.

facets() is used to query the cellxgene database for current values of one or all facets.

facets_filter() provides a convenient way to filter facets based on label or ontology term.

Usage

FACETS

facets(cellxgene_db = db(), facets = FACETS)

facets_filter(facet, key = c("label", "ontology_term_id"), value, exact = TRUE)

Format

FACETS is an object of class character of length 8.

Arguments

cellxgene_db

an (optional) cellxgene_db object, as returned by db().

facets

a character() vector corersponding to one of the facets in FACETS.

facet

the column containing faceted information, e.g., sex in datasets(db).

key

character(1) identifying whether value is a label or ontology_term_id.

value

character() value of the label or ontology term to filter on. The value may be a vector with length(value) > 0 for exact matchs (exact = TRUE, default), or a character(1) regular expression.

exact

logical(1) whether values match exactly (default, TRUE) or as a regular expression (FALSE).

Value

facets() returns a tibble with columns facet, label, ontology_term_id, and n, the number of times the facet label is used in the database.

facets_filter() returns a logical vector with length equal to the length (number of rows) of facet, with TRUE

indicating that the value of key is present in the dataset.

Examples

f <- facets()

## levels of each facet
f |>
    dplyr::count(facet)
#> # A tibble: 8 × 2
#>   facet                       n
#>   <chr>                   <int>
#> 1 assay                      36
#> 2 cell_type                 738
#> 3 development_stage         229
#> 4 disease                    91
#> 5 organism                    7
#> 6 self_reported_ethnicity    32
#> 7 sex                         3
#> 8 tissue                    373

## same as facets(, facets = "organism")
f |>
    dplyr::filter(facet == "organism")
#> # A tibble: 7 × 4
#>   facet    label                 ontology_term_id     n
#>   <chr>    <chr>                 <chr>            <int>
#> 1 organism Homo sapiens          NCBITaxon:9606     823
#> 2 organism Mus musculus          NCBITaxon:10090    265
#> 3 organism Callithrix jacchus    NCBITaxon:9483      27
#> 4 organism Macaca mulatta        NCBITaxon:9544      18
#> 5 organism Sus scrofa domesticus NCBITaxon:9825       3
#> 6 organism Gorilla gorilla       NCBITaxon:9593       1
#> 7 organism Pan troglodytes       NCBITaxon:9598       1

db <- db()
ds <- datasets(db)

## datasets with African American females
ds |>
    dplyr::filter(
        facets_filter(self_reported_ethnicity, "label", "African American"),
        facets_filter(sex, "label", "female")
    )
#> # A tibble: 50 × 25
#>    dataset_id   dataset_version_id collection_id donor_id assay  batch_condition
#>    <chr>        <chr>              <chr>         <list>   <list> <list>         
#>  1 66d15835-5d… 3b2b24ca-523e-438… a238e9fa-2bd… <list>   <list> <lgl [1]>      
#>  2 e47c65a8-7d… 27b49e7c-0d90-42c… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  3 c8d40d53-38… 48450fdc-ff62-498… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  4 c45f05b3-aa… d71044a8-2326-4ec… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  5 a6388a6f-60… 3cc5cb70-7442-4f5… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  6 a41202e6-17… 38b93dd0-5edd-47a… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  7 87183784-5b… ad925422-3011-470… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  8 842c6f5d-4a… 29cf079b-e74c-415… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  9 7f08cbcc-57… cd74e95e-6583-487… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#> 10 74520626-b0… 4e74cb9d-f8f5-43a… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#> # ℹ 40 more rows
#> # ℹ 19 more variables: cell_count <int>, cell_type <list>,
#> #   development_stage <list>, disease <list>, explorer_url <chr>,
#> #   is_primary_data <list>, mean_genes_per_cell <dbl>, organism <list>,
#> #   primary_cell_count <int>, schema_version <chr>,
#> #   self_reported_ethnicity <list>, sex <list>, suspension_type <list>,
#> #   tissue <list>, title <chr>, tombstone <lgl>, …

## datasets with non-European, known ethnicity
facets(db, "self_reported_ethnicity")
#> # A tibble: 32 × 4
#>    facet                   label                          ontology_term_id     n
#>    <chr>                   <chr>                          <chr>            <int>
#>  1 self_reported_ethnicity European                       HANCESTRO:0005     483
#>  2 self_reported_ethnicity unknown                        unknown            376
#>  3 self_reported_ethnicity na                             na                 308
#>  4 self_reported_ethnicity Asian                          HANCESTRO:0008     132
#>  5 self_reported_ethnicity African American               HANCESTRO:0568      59
#>  6 self_reported_ethnicity Hispanic or Latin American     HANCESTRO:0014      42
#>  7 self_reported_ethnicity admixed ancestry               HANCESTRO:0306      28
#>  8 self_reported_ethnicity African American or Afro-Cari… HANCESTRO:0016      26
#>  9 self_reported_ethnicity multiethnic                    multiethnic         25
#> 10 self_reported_ethnicity Greater Middle Eastern  (Midd… HANCESTRO:0015      22
#> # ℹ 22 more rows
ds |>
    dplyr::filter(
        !facets_filter(
            self_reported_ethnicity, "label", c("European", "na", "unknown")
        )
    )
#> # A tibble: 25 × 25
#>    dataset_id   dataset_version_id collection_id donor_id assay  batch_condition
#>    <chr>        <chr>              <chr>         <list>   <list> <list>         
#>  1 c45f05b3-aa… d71044a8-2326-4ec… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  2 87183784-5b… ad925422-3011-470… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  3 6e96d93c-de… 37e6e34e-ef55-4d8… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  4 31b49393-44… d6f77d6a-6a1e-4bf… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  5 2d3e3c41-34… 93c62ade-711e-49a… 4195ab4c-20b… <list>   <list> <lgl [1]>      
#>  6 cfa3c355-ee… 18f9edfa-8311-4d5… 9c8808ce-113… <list>   <list> <lgl [1]>      
#>  7 6ebb6e9b-ef… c546a784-39d1-43c… 8e880741-bf9… <list>   <list> <lgl [1]>      
#>  8 53ea4f06-03… 4bb26a38-0787-4aa… 8e880741-bf9… <list>   <list> <lgl [1]>      
#>  9 43600c3c-6a… c7819654-d26f-4cb… 8e880741-bf9… <list>   <list> <lgl [1]>      
#> 10 28ab6eb8-df… b0096575-4bbc-48a… 8e880741-bf9… <list>   <list> <lgl [1]>      
#> # ℹ 15 more rows
#> # ℹ 19 more variables: cell_count <int>, cell_type <list>,
#> #   development_stage <list>, disease <list>, explorer_url <chr>,
#> #   is_primary_data <list>, mean_genes_per_cell <dbl>, organism <list>,
#> #   primary_cell_count <int>, schema_version <chr>,
#> #   self_reported_ethnicity <list>, sex <list>, suspension_type <list>,
#> #   tissue <list>, title <chr>, tombstone <lgl>, …