Facets available for querying cellxgene data

FACETS is a character vector of common fields used to subset cellxgene data.

facets() is used to query the cellxgene database for current values of one or all facets.

facets_filter() provides a convenient way to filter facets based on label or ontology term.

Usage

FACETS

facets(cellxgene_db = db(), facets = FACETS)

facets_filter(facet, key = c("label", "ontology_term_id"), value, exact = TRUE)

Format

FACETS is an object of class character of length 8.

Arguments

cellxgene_db

an (optional) cellxgene_db object, as returned by db().

facets

a character() vector corersponding to one of the facets in FACETS.

facet

the column containing faceted information, e.g., sex in datasets(db).

key

character(1) identifying whether value is a label or ontology_term_id.

value

character() value of the label or ontology term to filter on. The value may be a vector with length(value) > 0 for exact matchs (exact = TRUE, default), or a character(1) regular expression.

exact

logical(1) whether values match exactly (default, TRUE) or as a regular expression (FALSE).

Value

facets() returns a tibble with columns facet, label, ontology_term_id, and n, the number of times the facet label is used in the database.

facets_filter() returns a logical vector with length equal to the length (number of rows) of facet, with TRUE

indicating that the value of key is present in the dataset.

Examples

f <- facets()

## levels of each facet
f |>
    dplyr::count(facet)
#> # A tibble: 8 × 2
#>   facet                       n
#>   <chr>                   <int>
#> 1 assay                      38
#> 2 cell_type                 773
#> 3 development_stage         232
#> 4 disease                   103
#> 5 organism                    7
#> 6 self_reported_ethnicity    32
#> 7 sex                         3
#> 8 tissue                    378

## same as facets(, facets = "organism")
f |>
    dplyr::filter(facet == "organism")
#> # A tibble: 7 × 4
#>   facet    label                 ontology_term_id     n
#>   <chr>    <chr>                 <chr>            <int>
#> 1 organism Homo sapiens          NCBITaxon:9606     866
#> 2 organism Mus musculus          NCBITaxon:10090    276
#> 3 organism Callithrix jacchus    NCBITaxon:9483      28
#> 4 organism Macaca mulatta        NCBITaxon:9544      19
#> 5 organism Sus scrofa domesticus NCBITaxon:9825       3
#> 6 organism Pan troglodytes       NCBITaxon:9598       2
#> 7 organism Gorilla gorilla       NCBITaxon:9593       1

db <- db()
ds <- datasets(db)

## datasets with African American females
ds |>
    dplyr::filter(
        facets_filter(self_reported_ethnicity, "label", "African American"),
        facets_filter(sex, "label", "female")
    )
#> # A tibble: 52 × 31
#>    dataset_id   dataset_version_id collection_id donor_id assay  batch_condition
#>    <chr>        <chr>              <chr>         <list>   <list> <list>         
#>  1 94c41723-b2… 4a9bcef2-02db-49f… f17b9205-f61… <list>   <list> <lgl [1]>      
#>  2 3de0ad6d-43… 6934232e-4db4-423… 625f6bf4-2f3… <list>   <list> <lgl [1]>      
#>  3 de985818-28… 64f14a2b-d754-4bc… c9706a92-0e5… <list>   <list> <lgl [1]>      
#>  4 0b4a15a7-4e… 6616739c-0b4a-4a4… a98b828a-622… <list>   <list> <lgl [1]>      
#>  5 3c361813-5d… 2e0b82ee-5320-437… e1fa9900-3fc… <list>   <list> <lgl [1]>      
#>  6 32b9bdce-24… 04019af8-aaf3-446… bcb61471-2a4… <list>   <list> <lgl [1]>      
#>  7 0b75c598-08… 6568e80d-b8d8-4a5… bcb61471-2a4… <list>   <list> <lgl [1]>      
#>  8 07854d9c-53… bf5607d5-82c3-46f… bcb61471-2a4… <list>   <list> <lgl [1]>      
#>  9 e40c6272-af… 23d7646e-5ed0-467… a48f5033-343… <list>   <list> <lgl [1]>      
#> 10 d6dfdef1-40… 5a2b8d31-f043-48e… a48f5033-343… <list>   <list> <lgl [1]>      
#> # ℹ 42 more rows
#> # ℹ 25 more variables: cell_count <int>, cell_type <list>, citation <chr>,
#> #   development_stage <list>, disease <list>, embeddings <list>,
#> #   explorer_url <chr>, feature_biotype <list>, feature_count <int>,
#> #   feature_reference <list>, is_primary_data <list>,
#> #   mean_genes_per_cell <dbl>, organism <list>, primary_cell_count <int>,
#> #   raw_data_location <chr>, schema_version <chr>, …

## datasets with non-European, known ethnicity
facets(db, "self_reported_ethnicity")
#> # A tibble: 32 × 4
#>    facet                   label                          ontology_term_id     n
#>    <chr>                   <chr>                          <chr>            <int>
#>  1 self_reported_ethnicity European                       HANCESTRO:0005     505
#>  2 self_reported_ethnicity unknown                        unknown            411
#>  3 self_reported_ethnicity na                             na                 320
#>  4 self_reported_ethnicity Asian                          HANCESTRO:0008     141
#>  5 self_reported_ethnicity African American               HANCESTRO:0568      61
#>  6 self_reported_ethnicity Hispanic or Latin American     HANCESTRO:0014      54
#>  7 self_reported_ethnicity Native American,Hispanic or L… HANCESTRO:0013,…    50
#>  8 self_reported_ethnicity African American or Afro-Cari… HANCESTRO:0016      32
#>  9 self_reported_ethnicity Greater Middle Eastern  (Midd… HANCESTRO:0015      22
#> 10 self_reported_ethnicity South Asian                    HANCESTRO:0006      11
#> # ℹ 22 more rows
ds |>
    dplyr::filter(
        !facets_filter(
            self_reported_ethnicity, "label", c("European", "na", "unknown")
        )
    )
#> # A tibble: 25 × 31
#>    dataset_id   dataset_version_id collection_id donor_id assay  batch_condition
#>    <chr>        <chr>              <chr>         <list>   <list> <list>         
#>  1 e6a11140-25… 4866a804-37eb-436… e5f58829-1a6… <list>   <list> <lgl [1]>      
#>  2 6ec405bb-47… c524a1cb-e823-478… e5f58829-1a6… <list>   <list> <lgl [1]>      
#>  3 2ba40233-85… a1e426d5-f575-445… e5f58829-1a6… <list>   <list> <lgl [1]>      
#>  4 2423ce2c-31… 80e4bc4f-e0a9-467… e5f58829-1a6… <list>   <list> <lgl [1]>      
#>  5 2adb1f8a-a6… 1b7484e3-83a0-47f… 38833785-fac… <list>   <list> <lgl [1]>      
#>  6 be884a28-0e… 7dd2aaf8-1714-49b… 10bf5c50-8d8… <list>   <list> <lgl [1]>      
#>  7 c8f83821-a2… 10ecac59-8895-492… 9d63fcf1-5ca… <list>   <list> <lgl [1]>      
#>  8 c05e6940-72… 80a4eb91-dab2-40d… 9d63fcf1-5ca… <list>   <list> <lgl [1]>      
#>  9 894573ad-49… e385cf86-3c0b-422… 9d63fcf1-5ca… <list>   <list> <lgl [1]>      
#> 10 84242d25-f6… 7ce8fb90-86fa-4ea… 9d63fcf1-5ca… <list>   <list> <lgl [1]>      
#> # ℹ 15 more rows
#> # ℹ 25 more variables: cell_count <int>, cell_type <list>, citation <chr>,
#> #   development_stage <list>, disease <list>, embeddings <list>,
#> #   explorer_url <chr>, feature_biotype <list>, feature_count <int>,
#> #   feature_reference <list>, is_primary_data <list>,
#> #   mean_genes_per_cell <dbl>, organism <list>, primary_cell_count <int>,
#> #   raw_data_location <chr>, schema_version <chr>, …