Query cellxgene collections, datasets, and files

files_download() retrieves one or more cellxgene files to a cache on the local system.

links(), authors() and publisher_metadata() are helper functions to extract 'nested' information from collections.

Usage

collections(cellxgene_db = db())

datasets(cellxgene_db = db())

datasets_visualize(tbl)

files(cellxgene_db = db())

files_download(tbl, dry.run = TRUE, cache.path = .cellxgene_cache_path())

links(cellxgene_db = db())

authors(cellxgene_db = db())

publisher_metadata(cellxgene_db = db())

Arguments

cellxgene_db

an optional 'cellxgene_db' object, as returned by db().

tbl

a tibble() typically derived from datasets(db) or files(db) and containing columns dataset_id (for datasets_visualize()), or columns dataset_id, file_id, and filetype (for files_download()).

dry.run

logical(1) indicating whether the (often large) file(s) in tbl should be downloaded to a local cache. Files are not downloaded when dry.run = TRUE (default).

cache.path

character(1) directory in which to cache downloaded files. The directory must already exist. The default is tools::R_user_dir("cellxgenedp", "cache"), a package-specific path in the user home directory.

Value

Each function returns a tibble describing the corresponding component of the database.

files_download() returns a character() vector of paths to the local files.

links() returns a tibble of external links associated with each collection. Common links includ DOI, raw data / data sources, and lab websites.

authors() returns a tibble of authors associated with each collection.

publisher_metadata() returns a tibble of publisher metadata (journal, publicate date, doi) associated with each collection.

Examples

db <- db()

collections(db)
#> # A tibble: 198 × 18
#>    collection_id    collection_version_id collection_url consortia contact_email
#>    <chr>            <chr>                 <chr>          <list>    <chr>        
#>  1 59c9ecfe-c47d-4… 8db59637-4b42-4195-b… https://cellx… <lgl [1]> twc@stanford…
#>  2 a18474f4-ff1e-4… e1078e5b-da88-49d8-b… https://cellx… <lgl [1]> kpn2114@colu…
#>  3 f5af7a2f-ab4c-4… cb80d978-a1a2-4696-8… https://cellx… <lgl [1]> Yanling_Liao…
#>  4 aee9c366-f2fb-4… 62fc92b9-92f9-4f6f-9… https://cellx… <lgl [1]> kjensen@tgen…
#>  5 c26ca66a-63ea-4… 9b5f598e-b5d6-4d83-a… https://cellx… <lgl [1]> semil.choksi…
#>  6 e5f58829-1a66-4… 519f5ac5-1f84-4b48-9… https://cellx… <chr [2]> angela.pisco…
#>  7 b52eb423-5d0d-4… e75822ca-72d5-4d7f-a… https://cellx… <chr [2]> st9@sanger.a…
#>  8 e3aa612b-0d7d-4… 8bfe4964-eab1-4ae8-b… https://cellx… <chr [1]> hongkuiz@all…
#>  9 1b014f39-f202-4… 8b7765f4-c81a-46fe-a… https://cellx… <lgl [1]> kimberly.ald…
#> 10 48d354f5-a5ca-4… be9483df-dbbd-421f-b… https://cellx… <chr [1]> Nathan.Salom…
#> # ℹ 188 more rows
#> # ℹ 13 more variables: contact_name <chr>, curator_name <chr>,
#> #   description <chr>, doi <chr>, links <list>, name <chr>,
#> #   publisher_metadata <list>, revising_in <lgl>, revision_of <lgl>,
#> #   visibility <chr>, created_at <date>, published_at <date>, revised_at <date>

collections(db) |>
    dplyr::glimpse()
#> Rows: 198
#> Columns: 18
#> $ collection_id         <chr> "59c9ecfe-c47d-4a6a-bab0-895cc0c1942b", "a18474f…
#> $ collection_version_id <chr> "8db59637-4b42-4195-bb8b-054cf7d03099", "e1078e5…
#> $ collection_url        <chr> "https://cellxgene.cziscience.com/collections/59…
#> $ consortia             <list> NA, NA, NA, NA, NA, <"CZ Biohub", "CZI Single-C…
#> $ contact_email         <chr> "twc@stanford.edu", "kpn2114@columbia.edu", "Yan…
#> $ contact_name          <chr> "Tony Wyss-Coray", "Karol Nowicki-Osuch", "Yanli…
#> $ curator_name          <chr> "Corinn Sophia Small", "Jason Hilton", "Jason Hi…
#> $ description           <chr> "Although SARS-CoV-2 primarily targets the respi…
#> $ doi                   <chr> "10.1038/s41586-021-03710-0", "10.1158/2159-8290…
#> $ links                 <list> [["GSE159812", "RAW_DATA", "https://www.ncbi.nl…
#> $ name                  <chr> "Dysregulation of brain and choroid plexus cell …
#> $ publisher_metadata    <list> [[["Yang", "Andrew C."], ["Kern", "Fabian"], ["…
#> $ revising_in           <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ revision_of           <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ visibility            <chr> "PUBLIC", "PUBLIC", "PUBLIC", "PUBLIC", "PUBLIC"…
#> $ created_at            <date> 2024-03-19, 2024-03-19, 2024-03-20, 2024-03-19,…
#> $ published_at          <date> 2024-02-07, 2023-03-15, 2024-02-15, 2024-01-16,…
#> $ revised_at            <date> 2024-03-22, 2024-03-22, 2024-03-22, 2024-03-22,…

datasets(db) |>
    dplyr::glimpse()
#> Rows: 1,285
#> Columns: 32
#> $ dataset_id                 <chr> "595c9010-99ec-462d-b6a1-2b2fe5407871", "2f…
#> $ dataset_version_id         <chr> "b4645848-e3d8-4920-bb4e-7aa294725177", "3b…
#> $ collection_id              <chr> "59c9ecfe-c47d-4a6a-bab0-895cc0c1942b", "59…
#> $ donor_id                   <list> <"15_12", "15_17", "16_18", "16_23", "16_2…
#> $ assay                      <list> [["10x 3' v3", "EFO:0009922"]], [["10x 3' …
#> $ batch_condition            <list> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "b…
#> $ cell_count                 <int> 27092, 38217, 3951, 146583, 5754, 3282, 293…
#> $ cell_type                  <list> [["cerebral cortex endothelial cell", "CL:…
#> $ citation                   <chr> "Publication: https://doi.org/10.1038/s4158…
#> $ default_embedding          <chr> "X_umap", "X_umap", NA, NA, NA, NA, NA, NA,…
#> $ development_stage          <list> [["55-year-old human stage", "HsapDv:00001…
#> $ disease                    <list> [["COVID-19", "MONDO:0100096"], ["aspirati…
#> $ embeddings                 <list> <"X_pca", "X_tsne", "X_umap">, <"X_pca", "…
#> $ explorer_url               <chr> "https://cellxgene.cziscience.com/e/595c901…
#> $ feature_biotype            <list> "gene", "gene", "gene", "gene", "gene", "g…
#> $ feature_count              <int> 36398, 36398, 33145, 33145, 33145, 33145, 2…
#> $ feature_reference          <list> "NCBITaxon:9606", "NCBITaxon:9606", "NCBIT…
#> $ is_primary_data            <list> TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, <T…
#> $ mean_genes_per_cell        <dbl> 1520.2394, 2091.8612, 1361.9355, 1830.6961,…
#> $ organism                   <list> [["Homo sapiens", "NCBITaxon:9606"]], [["H…
#> $ primary_cell_count         <int> 27092, 38217, 0, 0, 0, 0, 279586, 0, 0, 0, …
#> $ raw_data_location          <chr> "raw.X", "raw.X", "raw.X", "raw.X", "raw.X"…
#> $ schema_version             <chr> "5.0.0", "5.0.0", "5.0.0", "5.0.0", "5.0.0"…
#> $ self_reported_ethnicity    <list> [["European", "HANCESTRO:0005"]], [["Europ…
#> $ sex                        <list> [["female", "PATO:0000383"], ["male", "PAT…
#> $ suspension_type            <list> "nucleus", "nucleus", "cell", "cell", "cel…
#> $ tissue                     <list> [["choroid plexus", "UBERON:0001886", "tis…
#> $ title                      <chr> "A total of 27,092 droplet-based single-nuc…
#> $ tombstone                  <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, F…
#> $ x_approximate_distribution <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "CO…
#> $ published_at               <date> 2024-02-07, 2024-02-07, 2023-03-15, 2023-0…
#> $ revised_at                 <date> 2024-03-22, 2024-03-22, 2024-03-22, 2024-0…

# \donttest{
if (interactive()) {
    ## visualize the first dataset
    datasets(db) |>
        dplyr::slice(1) |>
        datasets_visualize()
}
# }
files(db) |>
    dplyr::glimpse()
#> Rows: 2,538
#> Columns: 4
#> $ dataset_id <chr> "595c9010-99ec-462d-b6a1-2b2fe5407871", "595c9010-99ec-462d…
#> $ filesize   <dbl> 333391557, 348066008, 630668977, 683935787, 31737910, 26350…
#> $ filetype   <chr> "H5AD", "RDS", "H5AD", "RDS", "H5AD", "RDS", "H5AD", "RDS",…
#> $ url        <chr> "https://datasets.cellxgene.cziscience.com/b4645848-e3d8-49…

if (FALSE) {
files(db) |>
    dplyr::slice(1) |>
    files_download(dry.run = FALSE)
}

## common links to external data
links(db) |>
    dplyr::count(link_type)
#> # A tibble: 5 × 2
#>   link_type       n
#>   <chr>       <int>
#> 1 DATA_SOURCE    36
#> 2 LAB_WEBSITE    42
#> 3 OTHER         351
#> 4 PROTOCOL       45
#> 5 RAW_DATA      302

## authors per collection
authors() |>
    dplyr::count(collection_id, sort = TRUE)
#> # A tibble: 189 × 2
#>    collection_id                            n
#>    <chr>                                <int>
#>  1 e5f58829-1a66-40b5-a624-9046778e74f5   221
#>  2 8f126edf-5405-4731-8374-b5ce11f53e82   205
#>  3 bcb61471-2a44-4d00-a0af-ff085512674c   171
#>  4 4f586cb6-972b-4ef7-a4ef-3c3800a3c004   147
#>  5 0b9d8a04-bb9d-44da-aa27-705bb65b54eb   135
#>  6 367d95c0-0eb0-4dae-8276-9407239421ee   106
#>  7 6f6d381a-7701-4781-935c-db10d30de293    98
#>  8 1ca90a2d-2943-483d-b678-b809bf464c30    94
#>  9 0a839c4b-10d0-4d64-9272-684c49a2c8ba    90
#> 10 ae1420fe-6630-46ed-8b3d-cc6056a66467    83
#> # ℹ 179 more rows

publisher_metadata() |>
    dplyr::glimpse()
#> Rows: 189
#> Columns: 9
#> $ collection_id   <chr> "59c9ecfe-c47d-4a6a-bab0-895cc0c1942b", "a18474f4-ff1e…
#> $ name            <chr> "Dysregulation of brain and choroid plexus cell types …
#> $ is_preprint     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE…
#> $ journal         <chr> "Nature", "Cancer Discovery", "Front. Immunol.", "Acta…
#> $ published_at    <date> 2021-07-22, 2023-03-17, 2023-09-20, 2023-09-01, 2022-…
#> $ published_year  <int> 2021, 2023, 2023, 2023, 2022, 2020, 2021, 2021, 2021, …
#> $ published_month <int> 7, 3, 9, 9, 5, 12, 6, 8, 3, 10, 12, 10, 9, 12, 12, 3, …
#> $ published_day   <int> 22, 17, 20, 1, 13, 17, 1, 1, 12, 7, 1, 14, 1, 1, 1, 20…
#> $ doi             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…