Query cellxgene collections, datasets, and files

files_download() retrieves one or more cellxgene files to a cache on the local system.

links(), authors() and publisher_metadata() are helper functions to extract 'nested' information from collections.

Usage

collections(cellxgene_db = db())

datasets(cellxgene_db = db())

datasets_visualize(tbl)

files(cellxgene_db = db())

files_download(tbl, dry.run = TRUE, cache.path = .cellxgene_cache_path())

links(cellxgene_db = db())

authors(cellxgene_db = db())

publisher_metadata(cellxgene_db = db())

Arguments

cellxgene_db

an optional 'cellxgene_db' object, as returned by db().

tbl

a tibble() typically derived from datasets(db) or files(db) and containing columns dataset_id (for datasets_visualize()), or columns dataset_id, file_id, and filetype (for files_download()).

dry.run

logical(1) indicating whether the (often large) file(s) in tbl should be downloaded to a local cache. Files are not downloaded when dry.run = TRUE (default).

cache.path

character(1) directory in which to cache downloaded files. The directory must already exist. The default is tools::R_user_dir("cellxgenedp", "cache"), a package-specific path in the user home directory.

Value

Each function returns a tibble describing the corresponding component of the database.

files_download() returns a character() vector of paths to the local files.

links() returns a tibble of external links associated with each collection. Common links includ DOI, raw data / data sources, and lab websites.

authors() returns a tibble of authors associated with each collection.

publisher_metadata() returns a tibble of publisher metadata (journal, publicate date, doi) associated with each collection.

Examples

db <- db()

collections(db)
#> # A tibble: 145 × 15
#>    collection_id   access_type consortia contact_email contact_name curator_name
#>    <chr>           <chr>       <list>    <chr>         <chr>        <chr>       
#>  1 2b02dff7-e427-… READ        <lgl [1]> miriam.merad… Miriam Merad Batuhan Cak…
#>  2 03cdc7f4-bd08-… READ        <lgl [1]> tien.peng@uc… Tien Peng    Jason Hilton
#>  3 a72afd53-ab92-… READ        <lgl [1]> cblish@stanf… Catherine A… Jennifer Yu…
#>  4 e9eec7f5-8519-… READ        <list>    mrc38@cam.ac… Menna R Cla… Batuhan Cak…
#>  5 4796c91c-9d8f-… READ        <lgl [1]> vazquezi@msk… Ignacio Váz… Jason Hilton
#>  6 bcb61471-2a44-… READ        <list>    info@kpmp.org KPMP         Jennifer Yu…
#>  7 13d1c580-4b17-… READ        <list>    my4@sanger.a… Matthew You… Jason Hilton
#>  8 5e469121-c203-… READ        <list>    biren@ucsd.e… Bing Ren     Jennifer Yu…
#>  9 9132fae8-bdfe-… READ        <lgl [1]> zhuang@chemi… Xiaowei Zhu… Jennifer Yu…
#> 10 0a77d4c0-d5d0-… READ        <list>    fabian.theis… Fabian Theis Jennifer Yu…
#> # ℹ 135 more rows
#> # ℹ 9 more variables: data_submission_policy_version <chr>, description <chr>,
#> #   links <list>, name <chr>, publisher_metadata <list>, visibility <chr>,
#> #   created_at <date>, published_at <date>, updated_at <date>

collections(db) |>
    dplyr::glimpse()
#> Rows: 145
#> Columns: 15
#> $ collection_id                  <chr> "2b02dff7-e427-4cdc-96fb-c0f354c099aa",…
#> $ access_type                    <chr> "READ", "READ", "READ", "READ", "READ",…
#> $ consortia                      <list> NA, NA, NA, ["Wellcome HCA Strategic S…
#> $ contact_email                  <chr> "miriam.merad@mssm.edu", "tien.peng@ucs…
#> $ contact_name                   <chr> "Miriam Merad", "Tien Peng", "Catherine…
#> $ curator_name                   <chr> "Batuhan Cakir", "Jason Hilton", "Jenni…
#> $ data_submission_policy_version <chr> "1.0", "1.0", "1.0", "1.0", "1.0", "1.0…
#> $ description                    <chr> "Clinical benefits of cytokine blockade…
#> $ links                          <list> [["", "RAW_DATA", "https://www.ncbi.nl…
#> $ name                           <chr> "Single-Cell Analysis of Crohn’s Diseas…
#> $ publisher_metadata             <list> [[["Martin", "Jerome C."], ["Chang", "…
#> $ visibility                     <chr> "PUBLIC", "PUBLIC", "PUBLIC", "PUBLIC",…
#> $ created_at                     <date> 2023-03-21, 2023-06-22, 2023-03-21, 20…
#> $ published_at                   <date> 2022-04-26, 2023-01-28, 2020-09-28, 20…
#> $ updated_at                     <date> 2023-03-21, 2023-06-22, 2023-03-21, 20…

datasets(db) |>
    dplyr::glimpse()
#> Rows: 841
#> Columns: 26
#> $ dataset_id              <chr> "7a637398-175f-4cf2-b06a-6a494503196b", "9aca7…
#> $ collection_id           <chr> "2b02dff7-e427-4cdc-96fb-c0f354c099aa", "03cdc…
#> $ donor_id                <list> ["pat.5", "pat.6", "pat.7", "pat.8", "pat.10"…
#> $ assay                   <list> [["10x 3' v1", "EFO:0009901"], ["10x 3' v2", …
#> $ cell_count              <int> 32458, 35699, 18386, 3662, 44721, 130908, 2116…
#> $ cell_type               <list> [["B cell", "CL:0000236"], ["alpha-beta T cel…
#> $ dataset_deployments     <chr> "https://cellxgene.cziscience.com/e/36c867a7-b…
#> $ development_stage       <list> [["human early adulthood stage", "HsapDv:0000…
#> $ disease                 <list> [["Crohn ileitis", "MONDO:0000709"]], [["norm…
#> $ is_primary_data         <chr> "PRIMARY", "PRIMARY", "PRIMARY", "SECONDARY", …
#> $ is_valid                <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE…
#> $ mean_genes_per_cell     <dbl> 763.68282, 1922.05370, 2918.85603, 3945.77171,…
#> $ name                    <chr> "Ileum", "immune cells", "non-immune cells", "…
#> $ organism                <list> [["Homo sapiens", "NCBITaxon:9606"]], [["Homo…
#> $ processing_status       <list> [0, "UPLOADED", "7a637398-175f-4cf2-b06a-6a49…
#> $ published               <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE…
#> $ revision                <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ schema_version          <chr> "3.0.0", "3.0.0", "3.0.0", "3.0.0", "3.0.0", "…
#> $ self_reported_ethnicity <list> [["unknown", "unknown"]], [["European", "HANC…
#> $ sex                     <list> [["unknown", "unknown"]], [["female", "PATO:0…
#> $ suspension_type         <list> ["cell"], ["cell"], ["cell"], ["cell"], ["cel…
#> $ tissue                  <list> [["ileum", "UBERON:0002116"]], [["alveolus of…
#> $ tombstone               <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALS…
#> $ created_at              <date> 2022-04-13, 2022-09-14, 2023-06-22, 2023-06-2…
#> $ published_at            <date> 2022-04-26, 2023-01-28, 2023-01-28, 2023-01-2…
#> $ updated_at              <date> 2022-04-13, 2022-09-14, 2023-06-22, 2023-06-2…

# \donttest{
## visualize the first dataset
datasets(db) |>
    dplyr::slice(1) |>
    datasets_visualize()
# }
files(db) |>
    dplyr::glimpse()
#> Rows: 3,348
#> Columns: 8
#> $ file_id        <chr> "e9b93ee2-37a7-454d-b2ea-ee369e979386", "78e930cd-3b37-…
#> $ dataset_id     <chr> "7a637398-175f-4cf2-b06a-6a494503196b", "7a637398-175f-…
#> $ filename       <chr> "", "local.rds", "raw.h5ad", "local.h5ad", "raw.h5ad", …
#> $ filetype       <chr> "CXG", "RDS", "RAW_H5AD", "H5AD", "RAW_H5AD", "H5AD", "…
#> $ s3_uri         <chr> "s3://hosted-cellxgene-prod/e9b93ee2-37a7-454d-b2ea-ee3…
#> $ user_submitted <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, T…
#> $ created_at     <date> 1970-01-01, 1970-01-01, 1970-01-01, 1970-01-01, 1970-0…
#> $ updated_at     <date> 1970-01-01, 1970-01-01, 1970-01-01, 1970-01-01, 1970-0…

if (FALSE) {
files(db) |>
    dplyr::slice(1) |>
    files_download(dry.run = FALSE)
}

## common links to external data
links(db) |>
    dplyr::count(link_type)
#> # A tibble: 6 × 2
#>   link_type       n
#>   <chr>       <int>
#> 1 DATA_SOURCE    40
#> 2 DOI           138
#> 3 LAB_WEBSITE    33
#> 4 OTHER         254
#> 5 PROTOCOL       40
#> 6 RAW_DATA      193

## authors per collection
authors() |>
    dplyr::count(collection_id, sort = TRUE)
#> # A tibble: 138 × 2
#>    collection_id                            n
#>    <chr>                                <int>
#>  1 e5f58829-1a66-40b5-a624-9046778e74f5   221
#>  2 8f126edf-5405-4731-8374-b5ce11f53e82   205
#>  3 4f586cb6-972b-4ef7-a4ef-3c3800a3c004   147
#>  4 0b9d8a04-bb9d-44da-aa27-705bb65b54eb   135
#>  5 367d95c0-0eb0-4dae-8276-9407239421ee   106
#>  6 6f6d381a-7701-4781-935c-db10d30de293    98
#>  7 1ca90a2d-2943-483d-b678-b809bf464c30    94
#>  8 0a839c4b-10d0-4d64-9272-684c49a2c8ba    90
#>  9 ae1420fe-6630-46ed-8b3d-cc6056a66467    83
#> 10 ddfad306-714d-4cc0-9985-d9072820c530    69
#> # ℹ 128 more rows

publisher_metadata() |>
    dplyr::glimpse()
#> Rows: 138
#> Columns: 9
#> $ collection_id   <chr> "2b02dff7-e427-4cdc-96fb-c0f354c099aa", "03cdc7f4-bd08…
#> $ name            <chr> "Single-Cell Analysis of Crohn’s Disease Lesions Ident…
#> $ is_preprint     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE,…
#> $ journal         <chr> "Cell", "Immunity", "Nat Med", "Current Opinion in Imm…
#> $ published_at    <date> 2019-09-01, 2023-02-01, 2020-07-01, 2022-06-01, 2022-…
#> $ published_year  <int> 2019, 2023, 2020, 2022, 2022, 2021, 2021, 2021, 2020, …
#> $ published_month <int> 9, 2, 7, 6, 12, 7, 12, 10, 6, 9, 1, 10, 9, 9, 11, 5, 1…
#> $ published_day   <int> 1, 1, 1, 1, 14, 29, 1, 7, 5, 1, 1, 7, 27, 6, 1, 13, 1,…
#> $ doi             <chr> "https://doi.org/10.1016/j.cell.2019.08.008", "https:/…