Query cellxgene collections, datasets, and files

files_download() retrieves one or more cellxgene files to a cache on the local system.

links(), authors() and publisher_metadata() are helper functions to extract 'nested' information from collections.

Usage

collections(cellxgene_db = db())

datasets(cellxgene_db = db())

datasets_visualize(tbl)

files(cellxgene_db = db())

files_download(tbl, dry.run = TRUE, cache.path = .cellxgene_cache_path())

links(cellxgene_db = db())

authors(cellxgene_db = db())

publisher_metadata(cellxgene_db = db())

Arguments

cellxgene_db

an optional 'cellxgene_db' object, as returned by db().

tbl

a tibble() typically derived from datasets(db) or files(db) and containing columns dataset_id (for datasets_visualize()), or columns dataset_id, file_id, and filetype (for files_download()).

dry.run

logical(1) indicating whether the (often large) file(s) in tbl should be downloaded to a local cache. Files are not downloaded when dry.run = TRUE (default).

cache.path

character(1) directory in which to cache downloaded files. The directory must already exist. The default is tools::R_user_dir("cellxgenedp", "cache"), a package-specific path in the user home directory.

Value

Each function returns a tibble describing the corresponding component of the database.

files_download() returns a character() vector of paths to the local files.

links() returns a tibble of external links associated with each collection. Common links includ DOI, raw data / data sources, and lab websites.

authors() returns a tibble of authors associated with each collection.

publisher_metadata() returns a tibble of publisher metadata (journal, publicate date, doi) associated with each collection.

Examples

db <- db()

collections(db)
#> # A tibble: 152 × 18
#>    collection_id    collection_version_id collection_url consortia contact_email
#>    <chr>            <chr>                 <chr>          <list>    <chr>        
#>  1 e75342a8-0f3b-4… 2d569157-4335-40d6-a… https://cellx… <list>    nhuebner@mdc…
#>  2 661a402a-2a5a-4… 626b26f4-a84c-4f31-8… https://cellx… <lgl [1]> rv4@sanger.a…
#>  3 367d95c0-0eb0-4… 4216ddce-94c0-4fdc-9… https://cellx… <list>    edl@allenins…
#>  4 af893e86-8e9f-4… fc8a8009-02f0-4084-9… https://cellx… <list>    ruichen@bcm.…
#>  5 48d354f5-a5ca-4… 54caf53d-0a4d-4874-9… https://cellx… <list>    Nathan.Salom…
#>  6 793fdaec-5067-4… b4431833-4155-48d7-8… https://cellx… <list>    m.a.haniffa@…
#>  7 13d1c580-4b17-4… c7b93415-bf09-45df-9… https://cellx… <list>    my4@sanger.a…
#>  8 fbc5881f-1ee3-4… 4f2da30d-407b-4c94-8… https://cellx… <list>    Douglas.Stra…
#>  9 c114c20f-1ef4-4… 871e2180-9ac4-4025-8… https://cellx… <lgl [1]> shendure@uw.…
#> 10 c8565c6a-01a1-4… 07bfe4f4-61bc-463c-a… https://cellx… <list>    carmen.sando…
#> # ℹ 142 more rows
#> # ℹ 13 more variables: contact_name <chr>, curator_name <chr>,
#> #   description <chr>, doi <chr>, links <list>, name <chr>,
#> #   publisher_metadata <list>, revising_in <lgl>, revision_of <lgl>,
#> #   visibility <chr>, created_at <date>, published_at <date>, revised_at <date>

collections(db) |>
    dplyr::glimpse()
#> Rows: 152
#> Columns: 18
#> $ collection_id         <chr> "e75342a8-0f3b-4ec5-8ee1-245a23e0f7cb", "661a402…
#> $ collection_version_id <chr> "2d569157-4335-40d6-af31-14e130c3d8bd", "626b26f…
#> $ collection_url        <chr> "https://cellxgene.cziscience.com/collections/e7…
#> $ consortia             <list> ["CZI Single-Cell Biology"], NA, ["Allen Instit…
#> $ contact_email         <chr> "nhuebner@mdc-berlin.de", "rv4@sanger.ac.uk", "e…
#> $ contact_name          <chr> "Norbert Hübner", "Roser Vento-Tormo", "Ed S. Le…
#> $ curator_name          <chr> "Jennifer Yu-Sheng Chien", "Batuhan Cakir", "Jen…
#> $ description           <chr> "Pathogenic variants in genes that cause dilated…
#> $ doi                   <chr> "10.1126/science.abo1984", "10.1038/s41586-022-0…
#> $ links                 <list> [["", "OTHER", "https://github.com/heiniglab/DC…
#> $ name                  <chr> "Pathogenic variants damage cell composition and…
#> $ publisher_metadata    <list> [[["Reichart", "Daniel"], ["Lindberg", "Eric L.…
#> $ revising_in           <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ revision_of           <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ visibility            <chr> "PUBLIC", "PUBLIC", "PUBLIC", "PUBLIC", "PUBLIC"…
#> $ created_at            <date> 2023-08-22, 2023-08-22, 2023-08-22, 2023-08-22,…
#> $ published_at          <date> 2022-08-05, 2023-04-11, 2021-02-17, 2021-10-29,…
#> $ revised_at            <date> 2023-08-22, 2023-08-22, 2023-08-22, 2023-08-22,…

datasets(db) |>
    dplyr::glimpse()
#> Rows: 910
#> Columns: 24
#> $ dataset_id                 <chr> "f7995301-7551-4e1d-8396-ffe3c9497ace", "ed…
#> $ dataset_version_id         <chr> "0a4f9a00-6f75-4ff7-aeec-d975b2a983fc", "61…
#> $ collection_id              <chr> "e75342a8-0f3b-4ec5-8ee1-245a23e0f7cb", "e7…
#> $ donor_id                   <list> ["DP2", "H3", "DT4", "H5", "DO1", "H6", "H…
#> $ assay                      <list> [["10x 3' v2", "EFO:0009899"], ["10x 3' v3…
#> $ batch_condition            <list> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ cell_count                 <int> 311418, 142816, 2576, 59217, 7999, 881081, …
#> $ cell_type                  <list> [["cardiac muscle cell", "CL:0000746"]], […
#> $ development_stage          <list> [["adolescent stage", "HsapDv:0000086"], […
#> $ disease                    <list> [["arrhythmogenic right ventricular cardio…
#> $ explorer_url               <chr> "https://cellxgene.cziscience.com/e/f799530…
#> $ is_primary_data            <list> [FALSE], [FALSE], [FALSE], [FALSE], [FALSE…
#> $ mean_genes_per_cell        <dbl> 2180.2172, 1437.7487, 2090.2655, 214.9302, …
#> $ organism                   <list> [["Homo sapiens", "NCBITaxon:9606"]], [["H…
#> $ schema_version             <chr> "3.1.0", "3.1.0", "3.1.0", "3.1.0", "3.1.0"…
#> $ self_reported_ethnicity    <list> [["Asian", "HANCESTRO:0008"], ["European",…
#> $ sex                        <list> [["female", "PATO:0000383"], ["male", "PAT…
#> $ suspension_type            <list> ["nucleus"], ["nucleus"], ["nucleus"], ["n…
#> $ tissue                     <list> [["apex of heart", "UBERON:0002098"], ["he…
#> $ title                      <chr> "DCM/ACM heart cell atlas: Cardiomyocytes",…
#> $ tombstone                  <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, F…
#> $ x_approximate_distribution <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ published_at               <date> 2022-08-05, 2022-08-05, 2022-08-05, 2022-0…
#> $ revised_at                 <date> 2023-08-22, 2023-08-22, 2023-08-22, 2023-0…

# \donttest{
if (interactive()) {
    ## visualize the first dataset
    datasets(db) |>
        dplyr::slice(1) |>
        datasets_visualize()
}
# }
files(db) |>
    dplyr::glimpse()
#> Rows: 1,802
#> Columns: 4
#> $ dataset_id <chr> "f7995301-7551-4e1d-8396-ffe3c9497ace", "f7995301-7551-4e1d…
#> $ filesize   <dbl> 3255625301, 3234403317, 1010106545, 967955201, 35165722, 26…
#> $ filetype   <chr> "H5AD", "RDS", "H5AD", "RDS", "H5AD", "RDS", "H5AD", "RDS",…
#> $ url        <chr> "https://datasets.cellxgene.cziscience.com/0a4f9a00-6f75-4f…

if (FALSE) {
files(db) |>
    dplyr::slice(1) |>
    files_download(dry.run = FALSE)
}

## common links to external data
links(db) |>
    dplyr::count(link_type)
#> # A tibble: 5 × 2
#>   link_type       n
#>   <chr>       <int>
#> 1 DATA_SOURCE    51
#> 2 LAB_WEBSITE    33
#> 3 OTHER         264
#> 4 PROTOCOL       40
#> 5 RAW_DATA      203

## authors per collection
authors() |>
    dplyr::count(collection_id, sort = TRUE)
#> # A tibble: 146 × 2
#>    collection_id                            n
#>    <chr>                                <int>
#>  1 e5f58829-1a66-40b5-a624-9046778e74f5   221
#>  2 8f126edf-5405-4731-8374-b5ce11f53e82   205
#>  3 bcb61471-2a44-4d00-a0af-ff085512674c   171
#>  4 4f586cb6-972b-4ef7-a4ef-3c3800a3c004   147
#>  5 0b9d8a04-bb9d-44da-aa27-705bb65b54eb   135
#>  6 367d95c0-0eb0-4dae-8276-9407239421ee   106
#>  7 6f6d381a-7701-4781-935c-db10d30de293    98
#>  8 1ca90a2d-2943-483d-b678-b809bf464c30    94
#>  9 0a839c4b-10d0-4d64-9272-684c49a2c8ba    90
#> 10 ae1420fe-6630-46ed-8b3d-cc6056a66467    83
#> # ℹ 136 more rows

publisher_metadata() |>
    dplyr::glimpse()
#> Rows: 146
#> Columns: 9
#> $ collection_id   <chr> "e75342a8-0f3b-4ec5-8ee1-245a23e0f7cb", "661a402a-2a5a…
#> $ name            <chr> "Pathogenic variants damage cell composition and singl…
#> $ is_preprint     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE…
#> $ journal         <chr> "Science", "Nature", "Nature", "Cell Genomics", "Scien…
#> $ published_at    <date> 2022-08-05, 2022-07-21, 2021-10-07, 2023-06-01, 2021-…
#> $ published_year  <int> 2022, 2022, 2021, 2023, 2021, 2021, 2021, 2020, 2020, …
#> $ published_month <int> 8, 7, 10, 6, 3, 10, 12, 8, 11, 10, 4, 4, 11, 12, 3, 11…
#> $ published_day   <int> 5, 21, 7, 1, 12, 14, 1, 1, 13, 7, 21, 1, 26, 1, 1, 26,…
#> $ doi             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…