Query cellxgene collections, datasets, and files

files_download() retrieves one or more cellxgene files to a cache on the local system.

links(), authors() and publisher_metadata() are helper functions to extract 'nested' information from collections.

Usage

collections(cellxgene_db = db())

datasets(cellxgene_db = db())

datasets_visualize(tbl)

files(cellxgene_db = db())

files_download(tbl, dry.run = TRUE, cache.path = .cellxgene_cache_path())

links(cellxgene_db = db())

authors(cellxgene_db = db())

publisher_metadata(cellxgene_db = db())

Arguments

cellxgene_db

an optional 'cellxgene_db' object, as returned by db().

tbl

a tibble() typically derived from datasets(db) or files(db) and containing columns dataset_id (for datasets_visualize()), or columns dataset_id, file_id, and filetype (for files_download()).

dry.run

logical(1) indicating whether the (often large) file(s) in tbl should be downloaded to a local cache. Files are not downloaded when dry.run = TRUE (default).

cache.path

character(1) directory in which to cache downloaded files. The directory must already exist. The default is tools::R_user_dir("cellxgenedp", "cache"), a package-specific path in the user home directory.

Value

Each function returns a tibble describing the corresponding component of the database.

files_download() returns a character() vector of paths to the local files.

links() returns a tibble of external links associated with each collection. Common links includ DOI, raw data / data sources, and lab websites.

authors() returns a tibble of authors associated with each collection.

publisher_metadata() returns a tibble of publisher metadata (journal, publicate date, doi) associated with each collection.

Examples

db <- db()

collections(db)
#> # A tibble: 183 × 18
#>    collection_id    collection_version_id collection_url consortia contact_email
#>    <chr>            <chr>                 <chr>          <list>    <chr>        
#>  1 ceb895f4-ff9f-4… ee098b5a-4f33-473b-b… https://cellx… <list>    panagiotis.r…
#>  2 af893e86-8e9f-4… 768170a6-c590-4900-a… https://cellx… <list>    ruichen@bcm.…
#>  3 1d1c7275-476a-4… 609becde-c797-41bb-8… https://cellx… <list>    wey334@g.har…
#>  4 1b014f39-f202-4… 1d88cb46-6e84-4b5b-b… https://cellx… <lgl [1]> kimberly.ald…
#>  5 48d354f5-a5ca-4… 2862daa3-c933-43c8-9… https://cellx… <list>    Nathan.Salom…
#>  6 43d4bb39-21af-4… 78360f02-1acc-415c-a… https://cellx… <lgl [1]> raymond.cho@…
#>  7 f7cecffa-00b4-4… 43224f82-db2a-443c-9… https://cellx… <list>    st9@sanger.a…
#>  8 f17b9205-f61f-4… 21ff4724-95e2-491b-8… https://cellx… <list>    genevieve.ko…
#>  9 64b24fda-6591-4… e414854b-2666-4977-9… https://cellx… <lgl [1]> magness@med.…
#> 10 48259aa8-f168-4… 44601b80-bd11-49d8-a… https://cellx… <lgl [1]> wtk22@cam.ac…
#> # ℹ 173 more rows
#> # ℹ 13 more variables: contact_name <chr>, curator_name <chr>,
#> #   description <chr>, doi <chr>, links <list>, name <chr>,
#> #   publisher_metadata <list>, revising_in <lgl>, revision_of <lgl>,
#> #   visibility <chr>, created_at <date>, published_at <date>, revised_at <date>

collections(db) |>
    dplyr::glimpse()
#> Rows: 183
#> Columns: 18
#> $ collection_id         <chr> "ceb895f4-ff9f-403a-b7c3-187a9657ac2c", "af893e8…
#> $ collection_version_id <chr> "ee098b5a-4f33-473b-b52a-0451de1f80ae", "768170a…
#> $ collection_url        <chr> "https://cellxgene.cziscience.com/collections/ce…
#> $ consortia             <list> ["BRAIN Initiative"], ["CZI Single-Cell Biology…
#> $ contact_email         <chr> "panagiotis.roussos@mssm.edu", "ruichen@bcm.edu"…
#> $ contact_name          <chr> "Panos Roussos", "Rui Chen", "Wenjun Yan", "Kimb…
#> $ curator_name          <chr> "Corinn Sophia Small", "Jennifer Yu-Sheng Chien"…
#> $ description           <chr> "We simultaneously profiled gene expression and …
#> $ doi                   <chr> "10.1126/sciadv.adg3754", "10.1016/j.xgen.2023.1…
#> $ links                 <list> [["SCP1859", "OTHER", "https://singlecell.broad…
#> $ name                  <chr> "Multi-omic profiling of the developing human ce…
#> $ publisher_metadata    <list> [[["Zhu", "Kaiyi"], ["Bendl", "Jaroslav"], ["Ra…
#> $ revising_in           <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ revision_of           <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ visibility            <chr> "PUBLIC", "PUBLIC", "PUBLIC", "PUBLIC", "PUBLIC"…
#> $ created_at            <date> 2023-12-12, 2023-12-11, 2023-12-12, 2023-12-11,…
#> $ published_at          <date> 2023-08-28, 2021-10-29, 2023-02-10, 2023-08-31,…
#> $ revised_at            <date> 2023-12-13, 2023-12-13, 2023-12-13, 2023-12-13,…

datasets(db) |>
    dplyr::glimpse()
#> Rows: 1,179
#> Columns: 31
#> $ dataset_id                 <chr> "53ce2631-3646-4172-bbd9-38b0a44d8214", "1d…
#> $ dataset_version_id         <chr> "2f17c183-388a-4c08-9adb-a146833e57ab", "94…
#> $ collection_id              <chr> "ceb895f4-ff9f-403a-b7c3-187a9657ac2c", "ce…
#> $ donor_id                   <list> ["LaFet1", "LaFet2", "EaFet2", "EaFet1", "…
#> $ assay                      <list> [["10x multiome", "EFO:0030059"]], [["10x …
#> $ batch_condition            <list> ["donor_id", "batch"], ["donor_id", "batch…
#> $ cell_count                 <int> 45549, 45549, 18011, 11617, 53040, 56507, 7…
#> $ cell_type                  <list> [["astrocyte", "CL:0000127"], ["caudal gan…
#> $ citation                   <chr> "Publication: https://doi.org/10.1126/sciad…
#> $ development_stage          <list> [["14-year-old human stage", "HsapDv:00001…
#> $ disease                    <list> [["normal", "PATO:0000461"]], [["normal", …
#> $ embeddings                 <list> ["X_joint_wnn_umap", "X_umap"], ["X_joint_…
#> $ explorer_url               <chr> "https://cellxgene.cziscience.com/e/53ce263…
#> $ feature_biotype            <list> ["gene"], ["gene"], ["gene"], ["gene"], ["…
#> $ feature_count              <int> 30113, 19492, 30933, 30933, 30933, 30933, 3…
#> $ feature_reference          <list> ["NCBITaxon:9606"], ["NCBITaxon:9606"], ["…
#> $ is_primary_data            <list> [TRUE], [TRUE], [FALSE], [TRUE], [FALSE], …
#> $ mean_genes_per_cell        <dbl> 1886.041, 4865.284, 2028.297, 4518.387, 222…
#> $ organism                   <list> [["Homo sapiens", "NCBITaxon:9606"]], [["H…
#> $ primary_cell_count         <int> 45549, 45549, 0, 11617, 0, 0, 0, 244474, 0,…
#> $ raw_data_location          <chr> "raw.X", "raw.X", "raw.X", "raw.X", "raw.X"…
#> $ schema_version             <chr> "4.0.0", "4.0.0", "4.0.0", "4.0.0", "4.0.0"…
#> $ self_reported_ethnicity    <list> [["unknown", "unknown"]], [["unknown", "un…
#> $ sex                        <list> [["female", "PATO:0000383"], ["male", "PAT…
#> $ suspension_type            <list> ["nucleus"], ["nucleus"], ["nucleus"], ["n…
#> $ tissue                     <list> [["cortical plate", "UBERON:0005343", "tis…
#> $ title                      <chr> "10x scRNA-seq from human cortex", "10x scA…
#> $ tombstone                  <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, F…
#> $ x_approximate_distribution <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ published_at               <date> 2023-08-28, 2023-08-28, 2021-10-29, 2021-1…
#> $ revised_at                 <date> 2023-12-13, 2023-12-13, 2023-12-13, 2023-1…

# \donttest{
if (interactive()) {
    ## visualize the first dataset
    datasets(db) |>
        dplyr::slice(1) |>
        datasets_visualize()
}
# }
files(db) |>
    dplyr::glimpse()
#> Rows: 2,338
#> Columns: 4
#> $ dataset_id <chr> "53ce2631-3646-4172-bbd9-38b0a44d8214", "53ce2631-3646-4172…
#> $ filesize   <dbl> 406108808, 399752425, 906795740, 1060800682, 1071401902, 14…
#> $ filetype   <chr> "H5AD", "RDS", "H5AD", "RDS", "H5AD", "RDS", "H5AD", "RDS",…
#> $ url        <chr> "https://datasets.cellxgene.cziscience.com/2f17c183-388a-4c…

if (FALSE) {
files(db) |>
    dplyr::slice(1) |>
    files_download(dry.run = FALSE)
}

## common links to external data
links(db) |>
    dplyr::count(link_type)
#> # A tibble: 5 × 2
#>   link_type       n
#>   <chr>       <int>
#> 1 DATA_SOURCE    35
#> 2 LAB_WEBSITE    38
#> 3 OTHER         333
#> 4 PROTOCOL       44
#> 5 RAW_DATA      275

## authors per collection
authors() |>
    dplyr::count(collection_id, sort = TRUE)
#> # A tibble: 173 × 2
#>    collection_id                            n
#>    <chr>                                <int>
#>  1 e5f58829-1a66-40b5-a624-9046778e74f5   221
#>  2 8f126edf-5405-4731-8374-b5ce11f53e82   205
#>  3 bcb61471-2a44-4d00-a0af-ff085512674c   171
#>  4 4f586cb6-972b-4ef7-a4ef-3c3800a3c004   147
#>  5 0b9d8a04-bb9d-44da-aa27-705bb65b54eb   135
#>  6 367d95c0-0eb0-4dae-8276-9407239421ee   106
#>  7 6f6d381a-7701-4781-935c-db10d30de293    98
#>  8 1ca90a2d-2943-483d-b678-b809bf464c30    94
#>  9 0a839c4b-10d0-4d64-9272-684c49a2c8ba    90
#> 10 ae1420fe-6630-46ed-8b3d-cc6056a66467    83
#> # ℹ 163 more rows

publisher_metadata() |>
    dplyr::glimpse()
#> Rows: 173
#> Columns: 9
#> $ collection_id   <chr> "ceb895f4-ff9f-403a-b7c3-187a9657ac2c", "af893e86-8e9f…
#> $ name            <chr> "Multi-omic profiling of the developing human cerebral…
#> $ is_preprint     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE…
#> $ journal         <chr> "Sci. Adv.", "Cell Genomics", "Sci Rep", "Nat Neurosci…
#> $ published_at    <date> 2023-10-13, 2023-06-01, 2020-06-17, 2021-08-01, 2021-…
#> $ published_year  <int> 2023, 2023, 2020, 2021, 2021, 2018, 2022, 2021, 2022, …
#> $ published_month <int> 10, 6, 6, 8, 3, 10, 12, 7, 1, 4, 12, 6, 9, 8, 10, 2, 1…
#> $ published_day   <int> 13, 1, 17, 1, 12, 1, 1, 1, 1, 21, 1, 15, 6, 1, 1, 1, 1…
#> $ doi             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…